add multi tenant databases support

+++ src/main/java/com/davidphu/catalog/jpa/CatalogRepository.java
@@ -2,5 +2,5 @@ package com.davidphu.catalog.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CatalogRepository extends JpaRepository<CatalogEntity, Integer> {
+public interface CatalogRepository extends JpaRepository<CatalogEntity, Long> {
 }
